Refactor to use common stable maths. Separate Weighted/Stable/BasePools.
